Fix null-dereference if there is a missing query
Avoid adding a callback in the strategies for languages which do not
have a query defined.  If an injected language does not have a query
defined we do not want to set up a callback for that language, otherwise
we get a nil-dereference exception when we try to iterate through the
captures of the non-existing query.