Use per-language namespaces for highlights
Highlighting needs to be removed sometimes. Highlighting is applied per
language and we can only clear entire lines of text, so there is a
possibility that we might clear highlighting from a previous language if
there is more than one language on a line of text.

With per-language namespaces we avoid this problem. Namespaces are
created lazily, i.e. the first time we look up the namespace ID for a
language that namespace is created. Afterwards the cached value is
reused.