Fix highlighting after undo operation
If the user uses an external application to format a file and then
undoes the changes the highlighting as messed up.  I have observed that
an empty change set was delivered from Tree-sitter.  My hack injects an
entry into the changeset to re-highlight the entire buffer, which hides
the problem from sight.