Extend `has_extmarks_at` test assertion
The assertion now takes an extra parameter to check for a specific
highlight group.  By default any highlight group will match, but if one
is specified only extmarks with that particular highlight groups are
counted.

While I was at it I also noticed that the `vim.fn.filter` function was
used wrongly, it did not mutate the Lua list in place.  I had to replace
it with `vim.tbl_filter`.